refactor(ui): rehome typeahead utility
